+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/PlotTypes/SignalXY.cs
@@ -59,4 +59,61 @@ public class SignalXY : ICategory
             myPlot.Add.SignalXY(xs, ys);
         }
     }
+
+    public class SignalXYRenderIndexes : RecipeBase
+    {
+        public override string Name => "Partial SignalXY Rendering";
+        public override string Description => "Even if a SignalXY plot references a " +
+            "large array of data, rendering can be limited to a range of values. If set," +
+            "only the range of data between the minimum and maximum render indexes will be displayed.";
+
+        [Test]
+        public override void Execute()
+        {
+            double[] xs = Generate.Consecutive(1000);
+            double[] ys = Generate.RandomWalk(1000);
+
+            var sigAll = myPlot.Add.SignalXY(xs, ys);
+            sigAll.Label = "Full";
+            sigAll.Data.YOffset = 80;
+
+            var sigLeft = myPlot.Add.SignalXY(xs, ys);
+            sigLeft.Label = "Left";
+            sigLeft.Data.YOffset = 60;
+            sigLeft.Data.MaximumIndex = 700;
+
+            var sigRight = myPlot.Add.SignalXY(xs, ys);
+            sigRight.Label = "Right";
+            sigRight.Data.YOffset = 40;
+            sigRight.Data.MinimumIndex = 300;
+
+            var sigMid = myPlot.Add.SignalXY(xs, ys);
+            sigMid.Label = "Mid";
+            sigMid.Data.YOffset = 20;
+            sigMid.Data.MinimumIndex = 300;
+            sigMid.Data.MaximumIndex = 700;
+
+            myPlot.ShowLegend(Alignment.UpperRight);
+            myPlot.Axes.Margins(top: .5);
+        }
+    }
+
+    public class SignalXYOffset : RecipeBase
+    {
+        public override string Name => "SignalXY Offset";
+        public override string Description => "A fixed offset can be applied to SignalXY plots.";
+
+        [Test]
+        public override void Execute()
+        {
+            double[] xs = Generate.Consecutive(1000);
+            double[] ys = Generate.Sin(1000);
+
+            var sig1 = myPlot.Add.SignalXY(xs, ys);
+
+            var sig2 = myPlot.Add.SignalXY(xs, ys);
+            sig2.Data.XOffset = 250;
+            sig2.Data.YOffset = .5;
+        }
+    }
 }