+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fonts: Improved font detection for strings containing mixed-language characters (#3220, #3184, #2746) _Thanks @kl7107 and @prime167_
 * Controls: Add a Reset function for context menus (#3224) _Thanks @MCF_
 * Axes: Prevent exceptions when generating ticks for a DateTime axis with zero size (#3221) _Thanks @devbotas_
+* SignalXY: Added `MinimumIndex` and `MaximumIndex` for partial array rendering (#3227)
+* SignalXY: Added `OffsetX` and `OffsetY` for for applying a fixed offset in coordinate space (#3227)
 
 ## ScottPlot 5.0.17
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-16_